Move tests in rank processor unit test to query term unit test and nearest neighbor query node unit test.
