+++ searchlib/src/tests/query/streaming/CMakeLists.txt
@@ -7,8 +7,10 @@ vespa_add_executable(searchlib_query_streaming_gtest_runner_app TEST
     hit_iterator_test.cpp
     hit_iterator_pack_test.cpp
     near_test.cpp
+    nearest_neighbor_query_node_test.cpp
     phrase_query_node_test.cpp
     query_builder_test.cpp
+    query_term_test.cpp
     same_element_query_node_test.cpp
     DEPENDS
     searchlib_queryeval_test
