+++ streamingvisitors/CMakeLists.txt
@@ -26,7 +26,6 @@ vespa_define_module(
     src/tests/nearest_neighbor_field_searcher
     src/tests/query_term_filter_factory
     src/tests/querywrapper
-    src/tests/rank_processor
     src/tests/searcher
     src/tests/searchvisitor
     src/tests/textutil
